+++ scripts/18-software/categories/internet/teams.sh
@@ -2,4 +2,9 @@
 
 echo "       - Microsoft Teams..."
 
+LOCK=/tmp/workstation.lock.internet.teams
+[ -f ${LOCK} ] && return
+
 flatpak install -y --noninteractive flathub com.github.IsmaelMartinez.teams_for_linux > /dev/null 2>&1
+
+> ${LOCK}
